# Conflicts:
#	.gitignore
#	android/app/build.gradle
#	ios/APP.xcodeproj/project.pbxproj
#	ios/APP/Info.plist

+++ src/data-access/datacache-provider.js
@@ -1,7 +1,14 @@
 const Realm = require('realm');
+import { AsyncStorage } from "react-native";
 import realmModel from '@models/realm-models';
 module.exports = {
-    save(userId, key, value) {
+    save(userId, key, value, saveToLocalStorage) {
+        if (saveToLocalStorage) {
+            AsyncStorage.setItem(userId + "_" + key, JSON.stringify({
+                value: value
+            }));
+            return;
+        }
         const { DataString, Schemas, schemaVersion } = realmModel;
         var _value = JSON.stringify({ data: value });
         try {
@@ -34,9 +41,21 @@ module.exports = {
             console.log(error);
         }
     },
-    read(userId, key, callback) {
+    read(userId, key, callback, fromLocalStorage) {
         if (callback) {
             try {
+                if (fromLocalStorage) {
+                    AsyncStorage.getItem(userId + "_" + key).then(value => {
+                        if (value) {
+                            let json = JSON.parse(value);
+                            callback(json.value);
+                        }
+                        else {
+                            callback(undefined, { message: "not found" });
+                        }
+                    });
+                    return;
+                }
                 const { Schemas, DataString, schemaVersion } = realmModel;
                 Realm.open({
                     schema: Schemas,
@@ -72,39 +91,51 @@ module.exports = {
             }
         }
     },
-    readPromise(userId, key) {
+    readPromise(userId, key, fromLocalStorage) {
         return new Promise((resolve, reject) => {
             try {
-                const { Schemas, DataString, schemaVersion } = realmModel;
-                Realm.open({
-                    schema: Schemas,
-                    schemaVersion,
-                    migration: function (oldRealm, newRealm) {
-                        newRealm.deleteAll();
-                    }
-                }).then((realm) => {
-                    try {
-                        let _key = "DataString_" + userId + "_" + key;
-                        var data = realm.objects(DataString.name).filtered("key == '" + _key + "'");
-                        if (data && data.length > 0) {
-                            var _value = JSON.parse(data[0].value);
-                            if (_value) {
-                                resolve(_value.data);
-                            }
+                if (fromLocalStorage) {
+                    AsyncStorage.getItem(userId + "_" + key).then(value => {
+                        if (value) {
+                            let json = JSON.parse(value);
+                            resolve(json.value);
                         }
                         else {
                             reject({ message: "not found" });
                         }
-                    } catch (e) {
+                    });
+                } else {
+                    const { Schemas, DataString, schemaVersion } = realmModel;
+                    Realm.open({
+                        schema: Schemas,
+                        schemaVersion,
+                        migration: function (oldRealm, newRealm) {
+                            newRealm.deleteAll();
+                        }
+                    }).then((realm) => {
+                        try {
+                            let _key = "DataString_" + userId + "_" + key;
+                            var data = realm.objects(DataString.name).filtered("key == '" + _key + "'");
+                            if (data && data.length > 0) {
+                                var _value = JSON.parse(data[0].value);
+                                if (_value) {
+                                    resolve(_value.data);
+                                }
+                            }
+                            else {
+                                reject({ message: "not found" });
+                            }
+                        } catch (e) {
+                            reject({
+                                message: e ? JSON.stringify(e) : ""
+                            });
+                        }
+                    }).catch(e => {
                         reject({
                             message: e ? JSON.stringify(e) : ""
                         });
-                    }
-                }).catch(e => {
-                    reject({
-                        message: e ? JSON.stringify(e) : ""
                     });
-                });
+                }
             } catch (e) {
                 reject({
                     message: e ? JSON.stringify(e) : ""
